usePan: pointer capture
Without it, when you move your mouse outside the element with button
pressed and release it, the element does not get the pointerup event
and still considers the button pressed and then pans when mouse pointer
is simply moved over it.

This affects only mouse, as for touch events, capture is implicit, for
some reason.