Extension system: fix extension context state defaults handling
We want to write the default to the state dictionary.

Otherwise this caused problems. E.g. when used to add an extra field to
question edit dialog (see e.g. QuestionGroupsCE), if the user opens the
edit dialog and does not change the groups field, it will get deleted,
as the extension's assembleQuestion hooks sees an empty state dict.